Revert choose_season to his original form & Add function to get the Seasons in watch_media
This commit refactors the choose_season function, reverting it to a
simple Select prompt, as it was originally. Additionally, it adds the
parse_seasons function, which returns the seasons of the media. These
changes are useful for implementing the option to select a season in the
menu or for choosing the same season as before.